stt/requirements.txt
flask==2.3.3
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0